Back to 'float' for the shooting stuff

+++ Assets/Sample Assets/2D/Scripts/Agent7ControlUI.cs
@@ -16,8 +16,8 @@ public class Agent7ControlUI : MonoBehaviour
 	bool jumpPressed = false;
 	bool shootPressed = false;
 	
-	double lastShotTime = 0;
-	double shootingThreshold = 0.7;
+	float lastShotTime = 0f;
+	float shootingThreshold = 0.7f;
 
 	void Awake()
 	{
